pyproject.toml
[project]
name = "crispy-isp-detector"
version = "0.1.0"
description = "Distributed ISP quality monitor - track RTT, packet loss, jitter, traceroutes, and speed across collectors"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "speedtest-cli>=2.1.3",
]

[project.scripts]
crispy-isp-detector = "collector.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["collector"]
